Split Application class up with an abstract base.
