+++ misuzu.php
@@ -3,6 +3,8 @@ namespace Misuzu;
 
 require_once 'vendor/autoload.php';
 
-$app = Application::start(__DIR__ . '/config/config.ini');
-$app->debug(IO\Directory::exists(__DIR__ . '/vendor/phpunit/phpunit'));
+$app = Application::start(
+    __DIR__ . '/config/config.ini',
+    IO\Directory::exists(__DIR__ . '/vendor/phpunit/phpunit')
+);
 $app->startDatabase();
